refactor: 에러 세분화, 오류 수정 (#172)
* feat: 이메일, 비밀번호, 닉네임 정규식 추가

* feat: DBError 정의 및 잘못된 필드 수정

* feat: Auth, User Error 정의 및 회원탈퇴 구현

* feat: Article Error 정의

* feat: Trade Error 정의 및 loadFilteredTradeList 메서드 필드 수정

* feat: Banner 에러처리

* feat: Storage Error 정의

* feat: Channel, Message Error 정의

+++ ANBDModel/Sources/Domain/Usecase/CommentUsecase.swift
@@ -25,13 +25,22 @@ public struct DefaultCommentUsecase: CommentUsecase {
     
     /// articleID가 일치하는 Article에 Comment를 작성하는 메서드
     public func writeComment(articleID: String, comment: Comment) async throws {
+        if articleID.isEmpty {
+            throw ArticleError.invalidArticleIDField
+        }
+        
         try await commentRepository.createComment(articleID: articleID, comment: comment)
     }
     
     
     /// articleID가 일치하는 Article의 Comment 배열을 반환하는 메서드
     public func loadCommentList(articleID: String) async throws -> [Comment] {
-        try await commentRepository.readCommentList(articleID: articleID)
+        if articleID.isEmpty {
+            throw ArticleError.invalidArticleIDField
+        }
+        
+        let commentList = try await commentRepository.readCommentList(articleID: articleID)
+        return commentList
     }
     
     
@@ -43,6 +52,10 @@ public struct DefaultCommentUsecase: CommentUsecase {
     
     /// commentID가 일치하는 Comment를 삭제하는 메서드
     public func deleteComment(articleID: String, commentID: String) async throws {
+        if articleID.isEmpty {
+            throw ArticleError.invalidArticleIDField
+        }
+        
         try await commentRepository.deleteComment(articleID: articleID, commentID: commentID)
     }
